requirements.txt
absl-py==0.14.0
addict==2.4.0
albumentations==1.0.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
cycler==0.10.0
Cython==0.29.24
e==1.4.5
ensemble-boxes==1.0.7
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.40.0
idna==3.2
imageio==2.9.0
joblib==1.0.1
kaggle==1.5.12
kiwisolver==1.3.2
llvmlite==0.37.0
Markdown==3.3.4
matplotlib==3.4.3
mean-average-precision==2021.4.26.0
mmcv-full==1.3.14
mmdet @ file:///home/quochungto/personal_projects/siim/src/detection/mmdet/mmdetection
networkx==2.6.3
numba==0.54.0
numpy==1.20.3
oauthlib==3.1.1
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
packaging==21.0
pandas==1.3.3
pathlib==1.0.1
Pillow==8.3.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminaltables==3.1.0
text-unidecode==1.3
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
tifffile==2021.8.30
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
yapf==0.31.0
